requirements.txt
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
djangorestframework==3.8.2
httpie==0.9.9
idna==2.6
more-itertools==4.1.0
mysqlclient==1.3.12
pluggy==0.6.0
py==1.5.3
Pygments==2.2.0
pytest==3.5.1
pytest-django==3.2.1
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
